release: don't assess apparmor when mocking nil features values
The value nil is special, it means we have not probed anything yet.
When the initial state was nil, nil (kernel and parser features,
respectively) and are mocked to some values a and b. Upon restoring the
statet will be nil, nil followed by a call to assess, which will
notice that no cached values exist and will proceed to probe the real
system.

This defeats the purpuse of the mocking system, where original values
are restored.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>