interfaces: ignore errors from apparmor kernel probe
This unit tests is not mocking the apparmor environment so it may
encounter any environment, including one where apparmor is not
available. This happens to be true in the place where Travis runs
the unit test smoke test:

    FAIL: system_key_test.go:75: systemKeySuite.TestInterfaceWriteSystemKey

    system_key_test.go:83:
	c.Assert(err, IsNil)
    ... value *os.PathError = &os.PathError{Op:"open",
    Path:"/sys/kernel/security/apparmor/features", Err:0x2} ("open
    /sys/kernel/security/apparmor/features: no such file or directory")

Since this is not a part of the test, ignore that error.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>