packaging/ubuntu: build with golang 1.10
This patch changes build dependency for snapd to pull in at least go
1.10. This way we can get off the 1.6 bandwagon (which has stopped and
is not going anywhere) and get the benefit of new and exciting features
of go 1.10.

The real hard code compatibility is go 1.9, up from 1.6 but not fully up
to 1.10 because some distributions are still used with only 1.9
available.

This branch _may_ help in chasing a core18 bug related to socket and
seeding. Fingers crossed.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>