release: use locking around lazy intialized state
Some state data in the release package, namely AppArmor level/summary &
kernel/parser features as well as Seccomp feature set, is only initialized when
needed. Since there is a chance, this information could be accessed in
concurrently, we need to add some locking around the code that mutates the data.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>