tests/main/selinux-clean: leave a note about snap logs
Snap logs will trigger snapd to run journalct. On a SELinux system, this will
cause a domain transition to journalctl_t. However, the policy does not allow
journalctl to poke /proc data of pid 1, thus causing denials to appear in the
log.

Until this is fixed, all we can do is have a TODO note as a reminder.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>