run-checks: stop running HEAD of staticcheck
The run-checks script runs a go static analysis tool staticcheck. The
tool was always fetched from master via "go get". Recently we found that
it complains about unused code (which is fine) but also about cgo and a
lot of little things that need to be inspected one by one.

I initially started fixing all of the issues but we should not have
broken master for long periods so let's start by nuking staticcheck.
We need a way to run a released version (and switch to another released
version explicitly). Ideally we'd also run some sort of meta-checker
that includes staticcheck internally.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>