devicestate: add initial Remodel support
This initial PR adds a new devicestate.Remodel() function that
takes a new model assertion and creates a change that will move
the system from the current model to the new model.

In this version of the remodel support is limited to just adding
new required snaps and to change the kernel track.