many: allow snap-update-ns to write user mount profile
When snap-confine constructs a per-user mount namespace it delegates the
task to snap-update-ns invoked with a special option. The code in
snap-update-ns switches to the appropriate user ID while retaining some
capabilities required for mounting.

Before snap-update-ns was never saving the per-user mount profile so it
didn't need anything. Now it may start saving it (when the appropriate
feature is enabled) but it runs under a non-root UID. To allow the save
to really happen allow the file path pattern via apparmor template of
snap-update-ns and retain more capabilities to bypass filesystem checks.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>