snap: split alignment calculation and display for channels
This is a bit of an RFC - it splits the channel display into
a first data gathering phase and a subsequent display phase.

This makes it (IMHO) easier to read the indent code for rev
and size because how the tasks of gathering the channel data
and the task of calculating the indent are stronger separated.

I hope this is useful, happy to also see it just as a starting
point for further work (or maybe even as inspiration :)