tests: review/fix the autopkgtest failures in disco
* tests: skip proxy test if https{,s}_proxy is set

The tinyproxy in the proxy test uses socket.socket(AF_INET) to connect
o the target for https and on systems that do not allow direct
connections to the outside world 443 port this will fail.

* tests: skip network-retry if http{,s}_proxy is set

The network-retry test may fail when http_proxy is set because
 it breaks DNS for the test. However if the proxy is available
and defined via an IP instead of a DNS name the test will not
work because the DNS is bypassed.

* tests: skip snap-service-watchdog in autopkgtest

Skip autopkgtest as this test is timing dependent and ADT is often
very slow.

Also fixes a incorrect print in the watchdog service binary:

   lib/snaps/test-snapd-service-watchdog/bin/direct

which lead to misleading:
```
Jan 09 20:28:55 autopkgtest test-snapd-service-watchdog.direct-watchdog-ok[21555]: INFO:root:watchdog notification every 0s
output
```

* address review feedback (thanks to John)

* test: review feedback for network-retry (thanks to mborzecki)

* tests: do not run some tests in autopkgtest

We get timeouts in autopkgtest on amd64 currently so this disables
some of the slower running tests in this environment. The real
issue is that the cloud that runs the autopkgtest is slow dues to
overcommitment it seems. But this is a reasonable workaround for
now.

* tests: re-enable install-store in ADT, disable snapctl-services/nfs-support