cmd/snap-update-ns: close internal lock on error
This patch fixes an error where a lock would internally leak until the
snap-update-ns process quits in the case where we would either:

 - fail to freeze the freezer cgroup
 - fail to verify a lock was already locked

The actual error is harmless since if either of those fail we just quit
anyway. The tests for this are already covered in the bigger refactor
branch where the locking code is broken out to a dedicated method for
easier testing.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>