cmd/snap-update-ns: refactor of profile application
The system wide and per-user mount profiles share a lot of the
application logic: in both cases the system needs some kind of locking
done to avoid races, the current and desired profile needs to be loaded
from disk, the sequence of changes needs to be computed and applied,
lastly the updated current profile needs to be saved.

Up until recently the per-user mount profile was never stored or
updated, just applied to an ephemeral mount namespace. This property is
not changed by this patch yet. Everything is shifted into place to make
that possible with minimal changes later.

All of the logic that performs the mount namespace update outlined above
is moved to executeMountProfileUpdate which takes an interface that
abstracts the various elements of the process. The system and user
update logic is refactored to offer that functionality accordingly. This
decomposition also allows each of the elements to be easily tested in
isolation, without having to setup extensive mocking or write to disk
each time.

Much of the common aspects of the system and user update code is further
moved to commonProfileUpdate, offering single code for locking, loading
and saving profiles, computing and applying mount changes.

System and user module as well as the common application glue are tested
with new unit tests, focusing on the key aspects of the differentiation.
The code no longer mixes tests for how to compute profile update, how to
apply profile update with every little other aspect that was previously
tested before.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>