cmd/snap-update-ns: split user switching code
We will soon want to be able to switch to a specific user and group
identifiers, not just the real ones. This use case will be used when
snapd is performing interface-triggered mount namespace update for
per-user mount namespaces. It will then run snap-update-ns with more
arguments to indicate that per-user mount namespace is being updated and
specify the user ID of user in question.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>