cmd/snap-update-ns: manually implement isspace
I wasn't able to find any other cause but using the isspace macro,
outside of tests (which pass just fine), causes ./snap-update-ns -u 1000
to crash. The debugger points to issspace(*uid_text) being the culprit
despite it not being a NULL pointer. It is correctly pointing to the
string "1000" when the crash happens. I was unable to solve it with a
static function that calls isspace internally (which is probably a
macro). As such, implement isspace ourselves to avoid the crash.

Upstream bug report: https://github.com/golang/go/issues/29689
also contains a minimal reproducer https://github.com/zyga/go-bug-29689
which crashes on latest golang binary release.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>