tests: exclude some more slow tests from runs in autopkgtest
The i386 autopkgtest run failed in disco with a timeout error.
This PR removes some more tests from the set of tests run in
the autopkgtest backend to bring the time below the timeout.

Note that the autopkgtest environment on i386/amd64 is
exceedingly slow in the ADT cloud, most likely because of
overcommit of resources there.