tests: define new "tests/smoke" suite and use that for autopkgtests
Our autopkgtest test backend runs the full "main" suite right now.
This causes issue on slow architectures or slow clouds because our
test suite with >300 integration tests regularly hits the autopkgtes
timeout of 2h46min in the amd64/i386 cloud environment.

Given that this has caused quite a few delays already when doing
snapd SRUs and that we run the full suite all the time in spread
this PR splits the main suite into a smaller "smoke" suite that
will be run inside the autopkgtest environment and that only
tests a small subset of the most important integration points.

Once we have a faster autopkgtest cloud we can switch back to
running the full tests again.