snapshotstate: don't task.Log without the lock
currently snapshotsate's `Restore` will pass its `task.Logf` straight
into `backend.Restore`, which will then use that to log messages for
the user.

The problem is that `backend.Restore` is run without the lock held,
meaning that `task.Logf` will panic.

This change fixes that, an adds a test that covers the situation as
well.