interfaces/network-manager: call netplan
Let NM call netplan so the device does not need to be re-started to
make the changes effective. Note that not confining the call to netplan
does not introduce additional risks in principle, as anyway we have
access to /etc/netplan so we have full control on what will happen the
next time netplan runs - eventuall, on next reboot.