snapd: fix race in TestSanityFailGoesIntoDegradedMode test
Ensure the MockSanityCheck code runs at least *twice* to
avoid a race in the TestSanityFailGoesIntoDegradedMode test:

If we close the channel and this wakes up the "select"
in the test immediately and stops this go-routine, then the
check that the logbuf contains the error will fail.
By running this at least twice we know the error made
it to the log.