many: add simple performance measure tool mostly for firstboot
This PR adds a very simplisit "measure" module to the tree that
can be used to take measurements of things that take time.

The API is minimal, the intended use (for now) is:
```
func foo() {
   m := measure.New("things happening in foo")
   defer m.Done()
   ...
}
```
The PR adds some measures into the firstboot code but we need to
experiment a bit too see where else we need to measure.

With that there is a new `snap debug measures` command that can
be used to get the most recent measurements.