overlord/snapstate/backend: look at the snap before calling fontconfig cache tools
When a non-core snap is installed, call the fc-cache helpers before the snap is
made runnable in the system (i.e. before current symlinks are established). This
ensures that before there is a chance to run the application, the cache has been
updated.

For the core snap, fc-cache is delivered as part of the snap, and the helpers
need to be called after the snap has been symlinked. Otherwise when upgrading
from a revision that did not ship the helpers, the caches would not get updated.

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>