overlord/snapstate: use an ad-hoc error when no results
When a single action returned no results (typically because the store
was shedding load), we were handling it as a "snap not found" (when
installing) or as an "no updates available" (when refreshing).

This is bad because it can confuse the user (snaps suddenly
disappearing for install), but worse it can confuse snapd into
allowing a refresh to a non-existent channel.

This change makes it so that the user gets an error telling them to
try again later, instead. Only for the single action case; a
multi-refresh or -install retains the old behaviour (so automatic
refresh doesn't start erroring). This means multi-install or -refresh
can still be confusing, but we can dig into that one later.