snapstate: auto transition on experimental.snapd-snap=true
This PR will automatically transition users to use the snapd
snap when the setting core.experimental.snapd-snap is set to
true.

Once this has landed we can ask people in the forum to set
this option to get wider user testing of this new feature
before we enable it eventually for everyone.