interfaces/dbus: be less strict about alternations for well-known names
Allow use of _ in the latter bits of the well-known name to support
Transmission and other applications adhering to dbus name best
practices. Unfortunately, the apparmor syntax does not allow for regexs,
so adjust to allow:

name=###DBUS_NAME###{_,-}[1-9]{,[0-9_]}{,[0-9_]}{,[0-9_]}{,[0-9_]}{,[0-9_]}{,[0-9_]}{,[0-9_]}{,[0-9_]}{,[0-9]}{,_[1-9]{,[0-9_]}{,[0-9_]}{,[0-9_]}{,[0-9_]}{,[0-9_]}{,[0-9_]}{,[0-9_]}{,[0-9_]}{,[0-9]}},

This allows the old -123456 (with more digits) as well as _<unsigned
long>_<unsigned long>. This continues to be safe so long as
interface were expanded to support parallel installs, there is an
opportunity for the parallel install and the original install to overlap
with this rule).

References:
https://forum.snapcraft.io/t/check-for-d-bus-session-bus-name-is-too-restrictive/8401/5