tests: remove -o pipefail
The `-o pipefail` option can have some unexpected side-effects
which are counter-intuitive and hard to debug/understand:

With SIGPIPE and "echo | grep -q" there is a race condition. If
grep exists before echo is finished writing data then echo will
get a SIGPIPE and this makes MATCH fail (which is very obscure
to the user). We use "echo | grep" e.g. in check_journalctl_log
and this failed in the new debian-sid suite.