cmd/snap-confine: rename and document seccomp profile loader
The function that loads a seccomp profile for a given security tag
is now called sc_apply_seccomp_profile_for_security_tag. A profile may
be a seccomp BPF filter or the special string "@unrestricted\n". In
addition there is non-trivial behavior when the profile is absent
that was never mentioned in the documentation before.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>