add DBus methods for controlling brightness and dynamically detect backlights
The kernel supports many different backlights with differing names in
/sys/devices, so listing them in static rules is cumbersome and there is
a maintenance burden of keeping in sync with new devices. The kernel
will put a symlink under /sys/class/backlight for each backlight, so we
can dereference the symlink for the /sys/devices paths. This works well
for permanently attached devices like laptop screens or other primary
screens since they aren't expected to change (even if they did on
reboot, snapd would notice this and adjust the policy).

snapd will not notice a new hotplugged monitor (but would if it was
plugged in on boot). A future update could utilize snapd's hotplug
support.