interfaces: add system-backup interface
Add a backup interface suitable for use on core and classic systems. On
core, allow access to /writable and on classic, all but /dev, /proc and
/sys.

This is not considered a super-privileged interface since it grants
precisely what it advertises, but we'll likely want to be very choosy
about granting auto-connection to snaps.