data/systemd: helper service to hold refreshes when waking up from suspend
Add a helper service that will launch after the system wakes up from sleep and
hold snapd refreshes for 10 minutes.

This aims to address a problem when the system would wake up and start
refreshing snaps right away, what may cause some undesired load and make the
machine unresponsive.

The service is set up so that it becomes part of `sleep.target`, which is used
when the host is going to sleep via either suspend, hibernation or a combination
of both. The low level /sys/power/state manipulation is implemented by
corresponding systemd helpers. The helpers exit when the system wakes up. Using
After dependency ordering, the service should start only after relevant sleep
helpers finish, effectively making it run after the wakeup. See [1] for
discussion about systemd side of things.

The default is to hold refresh for 10 minutes, and there is no way to change it
right now. Some initial feedback in the forum is positive so far [2].

References:
1. https://github.com/systemd/systemd/issues/6364
2. https://forum.snapcraft.io/t/defer-snapd-refresh-on-wake-from-suspend/4943

Signed-off-by: Maciej Borzecki <maciej.zenon.borzecki@canonical.com>