interfaces/apparmor: deny inet/inet6 in snap-update-ns profile
Sometimes snap-update-ns triggers inet and inet6 stream denials. This is
related to LP: #1595993. The denials are harmless in the case of
snap-update-ns and we don't want to allow it network access anyway.
Importantly, this would conflict with nfsSnippet, but nfsSnippet is not
added to snap-update-ns profiles, so there is no problem.