interfaces: add network-manager-observe interface
NetworkManager offers a very large API that is wholly exposed via the
network-manager interface. Some applications (eg, QtNetworkBearer) use
NetworkManager to check for connectivity and therefore do not require
access to the entire API. Implement the network-manager-observe
interface that exposes the Get* and List methods required for verifying
connectivity. Since this grants access to things like MAC addresses and
previous networks, treat this like network-observe and require manual
connection. This interface does not grant access to NetworkManager's
GetSecrets() API.

On classic distro, this interface lets the application communicate with
the system NetworkManager using the specified APIs. On all snaps
systems, a providing snap interested in exposing this smaller API is
expected to slot this interface in addition to the fuller
network-manager interface, which contains all the permanent slot policy
needed to run as NetworkManager (ie, this interface is currently not
standalone as there are no known applications that only implement this
reduced NM API).

References:
https://forum.snapcraft.io/t/qt-network-bearer-networkmanager-access-extend-network-observe/8399