cmd/snap-confine: handle death of helper process
When SIGPIPE is delivered the snap-confine process it would involuntary
terminate. If snap-confine was a part of a startup chain of a systemd
service then systemd would consider the exit as _clean_, as documented
by systemd.service(5), even though the actual snap application was never
invoked. This patch corrects this in the following way.

When the snap-confine helper process terminates due to SIGARLM-based
timeout mechanism the parent would receive SIGCHLD followed by SIGPIPE
on the inevitable attempt to communicate with the helper.

The default disposition for those signals are:

    SIGCHLD: ignore, restart system calls
    SIGPIPE: terminate

When the helper terminates the SIGCHLD signal is delivered immediately.
When snap-confine main process is not waiting for a system call then
nothing special happens. If, however, it is currently waiting on read or
write that call would be interrupted and then *restarted*, due to
default settings for system call restart semantics on Linux.

While changes to SIGCHLD handling were prototyped and were apparently
functioning correctly they are not a part of this patch in order to
reduce possible fallout in case there's unintended interaction with the
many children of the main snap-confine process.

Instead only one change is made: the disposition of SIGPIPE is set to
SIG_IGN, and only in the scope where we are interacting with the helper
process and are possibly using pipes for communication.

The patch also contains a regression test, for easier review.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>