cmd/snap-confine: allow tools to be scripts
When snap-confine invokes external tools it uses a helper that executes
the tool using fexecve, that is, exec-via-a-fd. When the executed
program is a regular binary that the kernel loads directly everything is
okay. When the program is a script the file descriptor that is being
passed cannot be O_CLOEXEC.

While we don't use scripts for snap-update-ns or snap-discard-ns, our
integration test suite would benefit from that. To allow scripts perform
the following changes:

 - After fork, immediately before fexevce perform a read of the first
   two bytes of the program, if they are "#!" then use fcntl to clear
   the FD_CLOEXEC flag.
 - To allow the read to operate switch from using O_PATH to regular
   O_RDONLY file descriptor for the opened executable.

Note that apparmor profiles for snap-confine and snap-update-ns prevent
them from running scripts. For testing the profile must be adjusted
externally.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>