packaging: disable systemd environment generator on 18.04
The systemd environment generator we ship must be disabled on 18.04
because of the systemd bug #1771858. This was reported for us as
LP: 1814355. With the generator on 18.04 we get a PATH that does
not include "/usr/local/{,s}bin which will break existing systems.

This PR disables the generator on 18.04 (which is most unfortunate)
until we can fix systemd.

A slightly simpler version of this fix was uploaded to bionic as
2.37.1.1.