snap-confine: fix classic snaps for users with /var/lib/* homedirs
When we removed the quirks system in commit 1da9316 we removed the
following apparmor rule as well:
```
     /var/lib/** rw,
```

The unintended side-effect of this change is that we broke snaps
for users with homedirs in /var/lib/ like "postgresql", "jenkins"
and potentially more. They now get the error:
```
cannot create user data directory: /var/lib/postgresql/snap/wal-e/13: Permission denied
```

We added a special case for jenkins which unfortunately is not
enough. This PR makes the snap-confine rule more general which
will unbreak all classic snaps that use /var/lib/* as their
homedir. This will fix the wal-e snap which is classic and needs
to run as the postgresql user to perform its tasks.

We probably still need a followup that deals with confined snaps
like jenkins in a more general manner like the current targeted
jenkins fix.