snap-confine: remove special handling of /var/lib/jenkins
We added special handling of the jenkins user in dd3c2f4 in
snap-confine by optionally mounting /var/lib/jenkins into the
core snap. However this is only needed to make *confined* snaps
work for the jenkins user which *never* worked.

When trying to run a confined snap (like test-snapd-tools) as
the jenkins user with an older versions of snapd we get the
following message:
```
cannot create user data directory: /var/lib/jenkins/snap/test-snapd-tools/6: Read-only file system
```

So this PR reverts the special handling of jenkins in snap-confine
becuase:
a) its a very special case for something that did not work before
b) it breaks existing users, I got a (private) bugreport that
   people get:
   `cannot perform operation: mount --rbind /var/lib/jenkins`

Note that this does *not* affect classic snaps which keep working
with this change for the special users that have homes in /var/lib.