cmd/snap-confine: handle death of helper process (#6466)
When SIGPIPE is delivered the snap-confine process it would involuntary
terminate. If snap-confine was a part of a startup chain of a systemd
service then systemd would consider the exit as _clean_, as documented
by systemd.service(5), even though the actual snap application was never
invoked. This patch corrects this in the following way.

When the snap-confine helper process terminates due to SIGARLM-based
timeout mechanism the parent would receive SIGPIPE on the attempt to
communicate with the helper.

The disposition of SIGPIPE is set to SIG_IGN, and only in the scope where we
are interacting with the helper process and are possibly using pipes for
communication. This ensures that the helper dies for whatever reason the
parent can just deal with EPIPE (and exiting with a proper error code) instead
of dying because of SIGPIPE.

The patch also contains a regression test, for easier review:

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>

* interfaces/apparmor: allow unconfined to signal snap-update-ns

This patch adds a rule that allows sending signals to snap-update-ns
from the unconfined world. This matches similar rule on snap-confine.

Such signal may be sent by, for example systemd, when a snap service is
being terminated soon after snap installation, when the update namespace
process is still alive.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>