daemon, dirs, overlord/snapstate: it's not a cache
I've been calling it 'store info cache' but @pedronis rightly points
out I don't go fetch things if there's a cache miss, so it's not a
cache; it's justan auxiliary data store. So I'm calling it that now.

Reads nicer anyway because 'auxiliary' has less roles (it's less often
a noun, and nearly never a verb).