snapstate: restart into the snapd snap on classic
With the addition of the "core.experimental.snapd-snap" option
we support running the snapd snap on classic and on core systems
now as well. When the snapd snap is available it is prefered for
re-exec over the core snap. This PR ensures that snapd restarts
itself when a classic snap is installed.