dirs,overlord/snapstate: add Soft and Hard refresh checks
As a part of "refresh app awareness" feature snapd should be able to
check if a snap is busy just before starting the refresh process and
then re-check just before stopping services and performing link changes.

There are two checks defined as a part of this flow: a soft check
that filters out processes that belong to services and a hard check that
simply looks if there are any processes running (that associate with given
snap).

This patch introduces this logic without actually using it yet. The soft
refresh check takes a snap.Info to know about apps and hooks. It looks
at the freezer cgroup which tracks all processes in a given snap
(regardless of origin) and then looks at the per-service PIDs cgroup
that tracks per-service processes. If there are any non-service processes
around then a snap is deemed busy.

Supplemental logic inspects apps and hooks to provide a descriptive
error message. Even though the process is partially racy and can provide
false positives or false negatives (e.g. a snap being busy just stopped
being busy or a snap not being busy may have just started to be busy)
this is deliberate. There is no freezing or locking at this step.

The second check is a hard check. After stopping services we can run the
hard check that simply looks if there are any processes left.

This time there is no synchronization internally but externally
(in the update logic) snapd will stop services, grab a lock and only then
perform the hard check and unlink a snap. This will ensure that the
process is free of any races. The exact semantics of the second lock
is not defined yet but it is explicitly orthogonal of this code.

The lock may be either the existing per-snap lock or a new lock that
has no timeout associated with it. This second lock would stall
startup of any new snap apps or hooks until the refresh process
can link the new revision.

Signed-off-by: Zygmunt Krynicki <me@zygoon.pl>