overlord/snapstate: add some randomness to the catalog refresh
Without this change, once the catalog refresh was successful it was
done every 24 hours exactly. This meant that if other external factors
conspired to synchronise a large number of machines to do a catalog
refresh at the same time, we'd get a thundering herd every day.

This change adds a delta to spread things out.