wrappers: Add an X-SnapName field to desktop files
Allow the snap that provides a .desktop file to be determined by putting the
snap name into the .desktop file (using the X-SnapName key).

This is useful for tools that scan the available .desktop files and want to get
additional snap information for that application. The particular motivation
for this change is gnome-control-center now exposes per-application
configuration and needs to determine if the .desktop file is from a Snap,
Flatpak (uses X-Flatpak key) or traditional packaging system.