pkg/apparmor: scratch documentation of apparmor
We're currently growing support for apparmor 2.13+ and I was thinking
about how we structure the code related to that. I'm worried of the
growing complexity in interfaces/apparmor (and release/apparmor.go) and
would like to move that code to pkg/apparmor where it can be referenced,
as appropriate from original locations. In addition I'd like to clearly
separate old and new apparmor, using the chance to document how it
behaves in each version.

This patch is by no means complete but I'd like to use it to start a
discussion about moving the code out, gardening it a little bit and
dropping it here. I chose the directory pkg/apparmor as we can expect
similar treatment to seccomp code without adding a new top-level element
for each package and the pkg/foo naming scheme seems to be common in the
go ecosystem.

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>